pyproject.toml
[tool.poetry]
name = "reinforced-gnn"
version = "1.0.0"
description = "Reinforced Graph Neural Networks for Collaborative Filtering"

license = "MIT"

readme = "README.md"
repository = "https://github.com/lfwa/reinforced-gnn"
homepage = "https://github.com/lfwa/reinforced-gnn"

[tool.poetry.dependencies]
python = "~3.8"
comet-ml = "3.8.1"
numpy = "1.20.2"
pandas = "1.2.4"
pytorch-lightning = "1.2.1"
scikit-learn = "0.24.1"
scikit-surprise = "1.1.1"
torch = "1.7.1"

# cudatoolkit = "10.1"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
flake8 = "^3.9.2"
isort = "^5.9.1"
mypy = "^0.910"
pre-commit = "^2.13.0"
pydocstyle = "^6.1.1"
pylint = "^2.9.0"


[tool.black]
line-length = 100
